Commit #3

+++ Controllers/RentingController.cs
@@ -5,6 +5,7 @@ using Microsoft.EntityFrameworkCore;
 using Rentals.Context;
 using Rentals.Models.DatabaseModel;
 using Rentals.Models.InputModel;
+using Rentals.Models.OutputModel;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -21,13 +22,13 @@ namespace Rentals.Controllers
         {
             _context = context;
         }
-        
+
         // *** znamená testováno a funkční
 
 
         //Přidat Výpůjčku
-        [HttpPost("Renting")]
-        public async Task<ActionResult<Renting>> AddNewRenting(RentingRequest renting)
+        [HttpPost()]
+        public async Task<ActionResult<Renting>> AddNewRenting([FromBody] RentingRequest renting)
         {
             var error = 0;
             if (_context.Users.Any(x => x.OauthId == renting.Owner) && _context.Users.Any(x => x.OauthId == renting.Approver))
@@ -42,26 +43,16 @@ namespace Rentals.Controllers
                     ApproverId = approver.Id,
                     Note = renting.Note,
                 };
-
-                //Stav
-                if ((DateTime.Compare(renting.Start, DateTime.Now) > 0))
-                {
-                    newRenting.State = RentingState.WillStart;
-                }
-                else
-                {
-                    newRenting.State = RentingState.InProgress;
-                }
-
+                newRenting.State = RentingState.WillStart;
                 _context.Rentings.Add(newRenting);
-                //await _context.SaveChangesAsync();
+                await _context.SaveChangesAsync();
 
                 //Přidání itemů
                 foreach (var item in renting.Items)
                 {
                     if (_context.Items.Any(x => x.Id == item) && _context.Items.Find(item).State == ItemState.Available && _context.Items.Find(item).IsDeleted == false)
                     {
-                        var rentingItem = new RentingItem { ItemId = item, RentingId = newRenting.Id }; // dostane se sem id i bez řádku 57?
+                        var rentingItem = new RentingItem { ItemId = item, RentingId = newRenting.Id };
                         _context.RentingItems.Add(rentingItem);
 
                         //Vložení itemu do uživatelova inventáře
@@ -88,7 +79,7 @@ namespace Rentals.Controllers
                 else
                 {
                     //{error} itemů nelze vypůjčit
-                    return Forbid($"Vyskytlo se {error} chyb");
+                    return BadRequest($"Vyskytlo se {error} chyb");
                 }
             }
             else
@@ -97,39 +88,39 @@ namespace Rentals.Controllers
             }
         }
 
-        //Úprava výpůjčky
-        [HttpPut("Renting")]
-        public async Task<ActionResult<Renting>> ChangeRenting(ChangeRentingRequest request)
+        //Vrácení předmětů výpůjčky
+        [HttpPut()]
+        public async Task<ActionResult<Renting>> ChangeRenting([FromBody] ChangeRentingRequest request)
         {
             var errors = 0;
             if (RentingExists(request.Id))
             {
                 Renting renting = _context.Rentings.Find(request.Id);
-                renting.End = request.End;
-                renting.Note = request.Note;
-
                 if (request.ReturnedItems != null)
                 {
                     foreach (var item in request.ReturnedItems)
                     {
                         //Pokud se ve výpůjčce nachází - vrácení itemu
-                        if (_context.RentingItems.Any(x => x.RentingId == request.Id && x.ItemId == item) && _context.InventoryItems.Any(x => x.UserId == renting.OwnerId && x.ItemId == item))
+                        if (_context.RentingItems.Any(x => x.RentingId == request.Id && x.ItemId == item && x.Returned == false) && _context.InventoryItems.Any(x => x.UserId == renting.OwnerId && x.ItemId == item))
                         {
-                            _context.RentingItems.Remove(_context.RentingItems.Single(x => x.RentingId == request.Id && x.ItemId == item));
+                            var Ritem = _context.RentingItems.SingleOrDefault(x => x.RentingId == request.Id && x.ItemId == item);
+                            Ritem.Returned = true;
+                            _context.Entry(Ritem).State = EntityState.Modified;
 
                             //Odebrání itemu z uživatelova inventáře
                             _context.InventoryItems.Remove(_context.InventoryItems.Single(x => x.UserId == renting.OwnerId && x.ItemId == item));
-
-                            if (!_context.RentingItems.Any(x => x.RentingId == request.Id))
-                            {
-                                renting.State = RentingState.Ended;
-                                renting.End = DateTime.Now;
-                            }
 
                             //Změna stavu itemu
                             var itemToReturn = _context.Items.Find(item);
                             itemToReturn.State = ItemState.Available;
                             _context.Entry(itemToReturn).State = EntityState.Modified;
+
+                            //Ukončení pokud vše vráceno
+                            if (!_context.RentingItems.Any(x => x.RentingId == request.Id))
+                            {
+                                renting.State = RentingState.Ended;
+                                renting.End = DateTime.Now;
+                            }
                         }
                         else
                         {
@@ -145,7 +136,7 @@ namespace Rentals.Controllers
                 else
                 {
                     //Špatné itemy? počet: {errors}
-                    return Forbid($"Vyskytlo se {errors} chyb");
+                    return BadRequest($"Vyskytlo se {errors} chyb");
                 }
             }
             else
@@ -154,28 +145,22 @@ namespace Rentals.Controllers
             }
         }
 
-        //Zrušení výpůjčky
-        [HttpDelete("Renting/{id}")]
+        //Zrušení neuskutečněné výpůjčky
+        [HttpDelete("{id}")]
         public async Task<ActionResult<Renting>> CancelRenting(int id)
         {
             if (RentingExists(id))
             {
                 Renting renting = _context.Rentings.Find(id);
 
-                //Smazání záznamů - dostupné itemy
+                //Odebrání itemů
                 foreach (var item in _context.RentingItems.Include(i => i.Item).Where(x => x.RentingId == id))
                 {
-                    item.Item.State = ItemState.Available;
                     _context.Entry(item).State = EntityState.Modified;
                     _context.RentingItems.Remove(item);
-
-                    //odebrání z inventáře
-                    _context.InventoryItems.Remove(_context.InventoryItems.Single(y => y.ItemId == item.ItemId && y.UserId == renting.OwnerId));
                 }
 
-                renting.State = RentingState.Cancelled;
-                _context.Entry(renting).State = EntityState.Modified;
-
+                _context.Remove(renting);
                 await _context.SaveChangesAsync();
                 return Ok(renting);
             }
@@ -186,17 +171,17 @@ namespace Rentals.Controllers
         }
 
         //Všechny výpůjčky + filtrace dle stavu
-        [HttpGet("Renting")]
+        [HttpGet()]
         public async Task<ActionResult<IEnumerable<Renting>>> GetAllRentings(RentingState? state)
         {
             IEnumerable<Renting> rentings = Enumerable.Empty<Renting>();
             if (state != null)
             {
-                 rentings = _context.Rentings.Include(o => o.Owner).Where(x => x.State == state).AsEnumerable();
+                rentings = _context.Rentings.Include(o => o.Owner).Where(x => x.State == state).AsEnumerable();
             }
             else
             {
-                 rentings = _context.Rentings.Include(o => o.Owner).AsEnumerable();
+                rentings = _context.Rentings.Include(o => o.Owner).AsEnumerable();
             }
             return Ok(rentings);
         }
@@ -215,7 +200,44 @@ namespace Rentals.Controllers
             {
                 return NotFound("Tento uživatel neexistuje");
             }
-        }   
+        }
+
+        //Aktivovat výpůjčku
+        [HttpPut("Activate/{id}")]
+        public async Task<ActionResult<Renting>> ActivateRenting(int id)
+        {
+            if (RentingExists(id) && _context.Rentings.SingleOrDefault(x => x.Id == id).State == RentingState.WillStart)
+            {
+                Renting renting = _context.Rentings.Find(id);
+                renting.State = RentingState.InProgress;
+                _context.Entry(renting).State = EntityState.Modified;
+                await _context.SaveChangesAsync();
+
+                return Ok(renting);
+            }
+            else
+            {
+                return BadRequest("Tuto výpůjčku nelze aktivovat nebo neexistuje");
+            }
+        }
+
+        //Data výpůjček
+        [HttpGet("Dates")]
+        public async Task<ActionResult<List<DatesResponse>>> GetDates()
+        {
+            List<DatesResponse> dates = new();
+            foreach (var item in _context.Rentings.Include(y => y.Owner).Where(x => x.State == RentingState.WillStart || x.State == RentingState.InProgress))
+            {
+                dates.Add(new DatesResponse
+                {
+                    Start = item.Start,
+                    End = item.End,
+                    Owner = item.Owner.FullName
+                });
+            }
+
+            return Ok(dates);
+        }
 
         private bool RentingExists(int id)
         {
